Revert changes regarding creating one audio player for app life to address bug where sound is heavily delayed if app is idle for a while
