+++ src/events/interactionCreate.js
@@ -1,9 +1,8 @@
 const { createReadStream } = require('node:fs');
 const { Events, MessageFlags } = require('discord.js');
-const { createAudioResource, demuxProbe, getVoiceConnection } = require('@discordjs/voice');
+const { createAudioPlayer, createAudioResource, demuxProbe, getVoiceConnection, AudioPlayerStatus } = require('@discordjs/voice');
 const { channelId } = require('../config.json');
 const Constants = require('../constants.js');
-const audioPlayerManager = require('../audioPlayerSingleton.js');
 
 module.exports = {
 	name: Events.InteractionCreate,
@@ -36,7 +35,8 @@ module.exports = {
 					return createAudioResource(stream, { inputType: type });
 				};
 
-				if (!getVoiceConnection(interaction.guild.id)) {
+				const connection = getVoiceConnection(interaction.guild.id);
+				if (!connection) {
 					throw new Error('The bot is not currently connected to a voice channel.');
 				}
 
@@ -45,7 +45,21 @@ module.exports = {
 					throw new Error('You are not currently in the voice channel, you cannot play sounds at this time.');
 				}
 
-				const player = audioPlayerManager.getAudioPlayer();
+				console.debug('Creating audio player');
+				const player = createAudioPlayer();
+
+				player.on('stateChange', (oldState, newState) => {
+					console.log(`Audio player transitioned from the ${oldState.status} state to the ${newState.status} state`);
+					if (oldState.status == AudioPlayerStatus.Playing && newState.status == AudioPlayerStatus.Idle) {
+						console.debug('The player finished playing a sound, destroying now');
+						player.stop();
+					}
+				});
+				player.on('error', error => {
+					console.error(`Audio player error: ${error.message}`);
+				});
+
+				connection.subscribe(player);
 				const location = `${Constants.AssetsFolder}/${interaction.customId}`;
 				const resource = await probeAndCreateResource(createReadStream(location));
 				player.play(resource);
@@ -53,7 +67,7 @@ module.exports = {
 			}
 			catch (error) {
 				console.error(`Error when button was clicked: ${error}`);
-				await interaction.reply({ content: error.toString(), flags: MessageFlags.Ephemeral });
+				await interaction.reply({ content: 'There was an issue playing the sound. The bot needs to be connected to the voice channel. Please try again later.', flags: MessageFlags.Ephemeral });
 			}
 		}
 		else {
