+++ src/events/ready.js
@@ -1,7 +1,6 @@
 const { Events } = require('discord.js');
 const { channelId } = require('../config.json');
 const { joinVoiceChannel } = require('@discordjs/voice');
-const { getAudioPlayer } = require('../audioPlayerSingleton.js');
 
 module.exports = {
 	name: Events.ClientReady,
@@ -13,12 +12,11 @@ module.exports = {
 
 		if (numConnectedMembers > 0) {
 			console.log(`Members already are connected; connecting bot to channel ${channel.id}, guild ${channel.guild.id}`);
-			const channelConnection = joinVoiceChannel({
+			joinVoiceChannel({
 				channelId: channel.id,
 				guildId: channel.guildId,
 				adapterCreator: channel.guild.voiceAdapterCreator,
 			});
-			channelConnection.subscribe(getAudioPlayer());
 		}
 
 		console.log(`Bot logged in as ${client.user.tag}`);
